third_party/blink/web_tests/fast/dom/document-importNode-arguments-expected.txt

This page tests calling document.importNode.

PASS imported.childNodes.length is 0
PASS imported.textContent == '' is true
PASS imported.childNodes.length is 2
PASS imported.textContent == 'hello world' is true
PASS imported.childNodes.length is 0
PASS imported.textContent == '' is true
PASS successfullyParsed is true

TEST COMPLETE
